+++ src/frontend/tests.zig
@@ -1,7 +1,4 @@
-const mm0 = @import("mm0");
-
 comptime {
     _ = @import("./tests/root.zig");
     _ = @import("./tests/unpack.zig");
-    _ = mm0.DefOpsTests;
 }